+++ src/lang/fr_FR.lang.php
@@ -199,6 +199,7 @@ $sm_lang = array(
 		'sms_gateway_smsglobal' => 'SMSGlobal',
 		'sms_gateway_smsit' => 'Smsit',
 		'sms_gateway_freevoipdeal' => 'FreeVoipDeal',
+		'sms_gateway_nexmo' => 'Nexmo',
 		'sms_gateway_username' => 'Nom utilisateur de la passerelle',
 		'sms_gateway_password' => 'Mot de passe de la passerelle',
 		'sms_from' => 'SMS de l\'expéditeur',