+++ src/lang/tr_TR.lang.php
@@ -198,6 +198,7 @@ $sm_lang = array(
         'sms_gateway_smsglobal' => 'SMSGlobal',
         'sms_gateway_smsit' => 'Smsit',
         'sms_gateway_freevoipdeal' => 'FreeVoipDeal',
+	'sms_gateway_nexmo' => 'Nexmo',
         'sms_gateway_username' => 'Servis kullanıcı adı',
         'sms_gateway_password' => 'Servis şifresi',
         'sms_from' => 'Gönderen numarası',